+++ ckeditor/views.py
@@ -20,22 +20,18 @@ except ImportError:
         return fn
         
 THUMBNAIL_SIZE = (75, 75)
-    
+
+# Use Django Storage API
+from django.core.files.storage import get_storage_class
+storage_class = get_storage_class()
+storage = storage_class(location=settings.CKEDITOR_UPLOAD_PATH, base_url=settings.CKEDITOR_UPLOAD_URL)
+   
 def get_available_name(name):
     """
     Returns a filename that's free on the target storage system, and
     available for new content to be written to.
     """
-    dir_name, file_name = os.path.split(name)
-    file_root, file_ext = os.path.splitext(file_name)
-    # If the filename already exists, keep adding an underscore (before the
-    # file extension, if one exists) to the filename until the generated
-    # filename doesn't exist.
-    while os.path.exists(name):
-        file_root += '_'
-        # file_ext includes the dot.
-        name = os.path.join(dir_name, file_root + file_ext)
-    return name
+    return storage.get_available_name(name)
 
 def get_thumb_filename(file_name):
     """
@@ -44,7 +40,7 @@ def get_thumb_filename(file_name):
     return '%s_thumb%s' % os.path.splitext(file_name)
 
 def create_thumbnail(filename):
-    image = Image.open(filename)
+    image = Image.open(storage.open(filename))
         
     # Convert to RGB if necessary
     # Thanks to Limodou on DjangoSnippets.org
@@ -54,20 +50,14 @@ def create_thumbnail(filename):
        
     # scale and crop to thumbnail
     imagefit = ImageOps.fit(image, THUMBNAIL_SIZE, Image.ANTIALIAS)
-    imagefit.save(get_thumb_filename(filename))
+
+    thumb_filename = get_thumb_filename(filename)
+    imagefit.save(storage.open(thumb_filename, 'w'))
+
         
-def get_media_url(path):
-    """
-    Determine system file's media URL.
-    """
-    upload_prefix = getattr(settings, "CKEDITOR_UPLOAD_PREFIX", None)
-    if upload_prefix:
-        url = upload_prefix + path.replace(settings.CKEDITOR_UPLOAD_PATH, '')
-    else:
-        url = settings.MEDIA_URL + path.replace(settings.MEDIA_ROOT, '')
-   
-    # Remove any double slashes.
-    return url.replace('//', '/')
+def get_media_url(path):   
+    return storage.url(path)
+
 
 def get_upload_filename(upload_name, user):
     # If CKEDITOR_RESTRICT_BY_USER is True upload file to user specific path.
@@ -77,17 +67,17 @@ def get_upload_filename(upload_name, user):
         user_path = ''
 
     # Generate date based path to put uploaded file.
-    date_path = datetime.now().strftime('%Y/%m/%d')
+    date_path = datetime.now().strftime('%Y%m%d')
     
     # Complete upload path (upload_path + date_path).
-    upload_path = os.path.join(settings.CKEDITOR_UPLOAD_PATH, user_path, date_path)
-   
-    # Make sure upload_path exists.
-    if not os.path.exists(upload_path):
-        os.makedirs(upload_path)
-    
+    upload_path = os.path.join(
+            user_path,
+            date_path,
+            storage.get_valid_name(upload_name)
+            )
+      
     # Get available name and return.
-    return get_available_name(os.path.join(upload_path, upload_name))
+    return storage.get_available_name(upload_path)
      
     
 @csrf_exempt
@@ -100,21 +90,16 @@ def upload(request):
     """
     # Get the uploaded file from request.
     upload = request.FILES['upload']
-    upload_ext = os.path.splitext(upload.name)[1]
-   
-    # Open output file in which to store upload. 
-    upload_filename = get_upload_filename(upload.name, request.user)
-    out = open(upload_filename, 'wb+')
 
-    # Iterate through chunks and write to destination.
-    for chunk in upload.chunks():
-        out.write(chunk)
-    out.close()
+    upload_filename = get_upload_filename(upload.name, request.user)
+
+    storage.save(upload_filename, upload)
 
     create_thumbnail(upload_filename)
 
     # Respond with Javascript sending ckeditor upload url.
-    url = get_media_url(upload_filename)
+    url = storage.url(upload_filename)
+
     return HttpResponse("""
     <script type='text/javascript'>
         window.parent.CKEDITOR.tools.callFunction(%s, '%s');
@@ -138,13 +123,14 @@ def get_image_browse_urls(user=None):
     
     for root, dirs, files in os.walk(browse_path):
         for filename in [ os.path.join(root, x) for x in files ]:
+            filename = filename.replace(settings.CKEDITOR_UPLOAD_PATH,'')
             # bypass for thumbs
             if '_thumb' in filename:
                 continue
             
             images.append({
-                'thumb': get_media_url(get_thumb_filename(filename)),
-                'src': get_media_url(filename)
+                'thumb': storage.url(get_thumb_filename(filename)),
+                'src': storage.url(filename)
             })
 
     return images